requirements.txt
langchain
openai
sqlalchemy
pymysql
python-dotenv